requirements.txt
# Core data processing
pandas>=2.0.0

# HTTP requests for API calls
requests>=2.28.0

# OpenAI API for LLM pattern detection
openai>=1.0.0

# Environment variable management
python-dotenv>=1.0.0


